Use direct signing on gov prop creation.
